Move more functionality to useDataStoreGridBase composable
